Refactor and rename Github service
Renamed and separated the responsibilities of the Github service into separate modules to increase it's readability.
